Move TileRenderResult to its own class
