+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/config/characteristics/TileSource.kt
@@ -1,6 +1,6 @@
 package com.rafambn.kmap.config.characteristics
 
-import com.rafambn.kmap.model.TileRenderResult
+import com.rafambn.kmap.utils.TileRenderResult
 
 interface TileSource {
     suspend fun getTile(zoom: Int, row: Int, column: Int): TileRenderResult
