+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/core/state/TileCanvasState.kt
@@ -2,7 +2,7 @@ package com.rafambn.kmap.core.state
 
 import com.rafambn.kmap.core.TileLayers
 import com.rafambn.kmap.model.Tile
-import com.rafambn.kmap.model.TileRenderResult
+import com.rafambn.kmap.utils.TileRenderResult
 import com.rafambn.kmap.model.TileSpecs
 import com.rafambn.kmap.utils.loopInZoom
 import kotlinx.coroutines.CoroutineScope
