+++ KMaPDemo/src/commonMain/kotlin/com/rafambn/kmap/customSources/OSMTileSource.kt
@@ -3,7 +3,7 @@ package com.rafambn.kmap.customSources
 import androidx.compose.ui.graphics.ImageBitmap
 import com.rafambn.kmap.config.characteristics.TileSource
 import com.rafambn.kmap.model.Tile
-import com.rafambn.kmap.model.TileRenderResult
+import com.rafambn.kmap.utils.TileRenderResult
 import com.rafambn.kmap.model.TileSpecs
 import io.ktor.client.HttpClient
 import io.ktor.client.request.get
